Issue in examples/nestedGuards.real
For some reason the nestedGuards.real's `demo` stops consuming the CPU completely without
the call to f2, but with the call in the case statement, it does not.
The case statement specifically shows that a regular nested case
statement stops but if nesting is done through a function call, it does
not.